requirements.txt
zenml==0.84.3
dvc==3.30.1
hydra-core==1.3.2
optuna==3.4.0
evidently==0.5.0
mlflow==2.8.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
pytest==7.4.3
loguru==0.7.2
PyYAML==6.0.1
jupyter
fastapi
uvicorn
flask==2.3.3
pydantic==2.5.0
scipy==1.11.4
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
kneed==0.8.5
yellowbrick==1.5
shap==0.44.1
lime==0.2.0.1
eli5==0.13.0
mlxtend==0.23.1
imbalanced-learn==0.11.0
category_encoders==2.6.3
featuretools==1.28.0
tsfresh==0.20.1
statsmodels==0.14.0
plotly==5.17.0
kaleido==0.2.1
streamlit==1.28.1
gradio==4.7.1
bentoml==1.1.11
seldon-core==2.8.0
kserve==0.11.2
mlserver==1.3.5
alibi-detect==0.11.4
alibi-explain==0.9.4
great-expectations==0.18.8
pandera==0.18.0
deepchecks==0.18.1
# Security dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
bandit==1.7.6
safety==2.3.5
